requirements.txt
Django==3.2.13
djangorestframework==3.13.1
psycopg2-binary==2.9.3
daphne
whitenoise
pyyaml
uritemplate
channels
channels_redis
redis
celery
django-celery-beat
asgiref